+++ sprockets/mixins/metrics/__init__.py
@@ -1,3 +1,3 @@
-version_info = (4, 0, 0)
+version_info = (4, 1, 0)
 __version__ = '.'.join(str(v) for v in version_info)
 __all__ = ['__version__', 'version_info']